add support to to_singer to build schema from catalog
