+++ gluestick/__init__.py
@@ -3,3 +3,4 @@
 from .etl_utils import *  # noqa
 from .pandas_utils import *  # noqa
 from .singer import *  # noqa
+from .reader import *  # noqa
